Progess bar + description box
Description box lets user add a description, theres a getter in  the main scene to retrieve it and the list of tags from the tag box once user has finished. Progess bar added below the tabPane which shows the time of the scene, i.e maximum duration of elements.

+++ src/main/java/sweng/group/one/client_app_desktop/media/GraphicsElement.java
@@ -94,7 +94,7 @@ public class GraphicsElement extends PresElement{
 		component.repaint();
 	}
 	
-	public void setDuration(float duration) {
+	public void setDuration(long duration) {
 		this.duration=duration;
 	}
 }
